+++ resources/views/examples/quantity-selector/disabled.blade.php
@@ -1,1 +1,0 @@
-<x-ui.quantity-selector name="quantity" :value="3" :disabled="true" />