+++ resources/views/examples/quantity-selector/with-max.blade.php
@@ -1,1 +1,0 @@
-<x-ui.quantity-selector name="quantity" :value="2" :min="1" :max="10" />